MAINT: Refactor test into 2 separate files
